+++ lib/Controller/ApiController.php
@@ -196,6 +196,7 @@ class ApiController extends Controller {
 	public function getScanState() {
 		return new JSONResponse([
 			'unscannedFiles' => $this->scanner->getUnscannedMusicFileIds($this->userId),
+			'dirtyFiles' => $this->scanner->getDirtyMusicFileIds($this->userId),
 			'scannedCount' => $this->trackBusinessLayer->count($this->userId)
 		]);
 	}