+++ lib/Migration/Version020100Date20241114220300.php
@@ -1,0 +1,59 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OCA\Music\Migration;
+
+use Closure;
+use OCP\DB\ISchemaWrapper;
+use OCP\Migration\SimpleMigrationStep;
+use OCP\Migration\IOutput;
+
+/**
+ * Migrate the DB schema to Music v2.1.0 level from the v1.9.1 level
+ */
+class Version020100Date20241114220300 extends SimpleMigrationStep {
+
+	/**
+	 * @param IOutput $output
+	 * @param Closure $schemaClosure The `\Closure` returns a `ISchemaWrapper`
+	 * @param array $options
+	 */
+	public function preSchemaChange(IOutput $output, Closure $schemaClosure, array $options) {
+	}
+
+	/**
+	 * @param IOutput $output
+	 * @param Closure $schemaClosure The `\Closure` returns a `ISchemaWrapper`
+	 * @param array $options
+	 * @return null|ISchemaWrapper
+	 */
+	public function changeSchema(IOutput $output, Closure $schemaClosure, array $options) {
+		/** @var ISchemaWrapper $schema */
+		$schema = $schemaClosure();
+		$this->addDirtyFieldToTrack($schema);
+		return $schema;
+	}
+
+	/**
+	 * @param IOutput $output
+	 * @param Closure $schemaClosure The `\Closure` returns a `ISchemaWrapper`
+	 * @param array $options
+	 */
+	public function postSchemaChange(IOutput $output, Closure $schemaClosure, array $options) {
+	}
+
+	/**
+	 * Add the new field 'dirty' to the table 'music_tracks'
+	 */
+	private function addDirtyFieldToTrack(ISchemaWrapper $schema) {
+		$table = $schema->getTable('music_tracks');
+		$this->setColumn($table, 'dirty', 'smallint', ['notnull' => true, 'default' => 0]);
+	}
+
+	private function setColumn($table, string $name, string $type, array $args) : void {
+		if (!$table->hasColumn($name)) {
+			$table->addColumn($name, $type, $args);
+		}
+	}
+}