+++ templates/main.php
@@ -48,11 +48,19 @@ HtmlUtil::addWebpackStyle('app');
 				</div>
 			</div>
 
-			<div id="toScan" class="emptycontent clickable" ng-show="toScan && viewingLibrary()" ng-click="startScanning()">
+			<div id="toScan" class="emptycontent clickable" ng-show="!scanning && unscannedFiles.length && viewingLibrary()" ng-click="startScanning(unscannedFiles)">
 				<div class="icon-audio svg"></div>
 				<div>
 					<h2 translate>New music available</h2>
 					<p translate>Click here to start the scan</p>
+				</div>
+			</div>
+
+			<div id="toRescan" class="emptycontent clickable" ng-show="!scanning && !unscannedFiles.length && dirtyFiles.length && viewingLibrary()" ng-click="startScanning(dirtyFiles)">
+				<div class="icon-audio svg"></div>
+				<div>
+					<h2 translate>Some of the previously scanned files may have changed</h2>
+					<p translate>Click here to rescan these files</p>
 				</div>
 			</div>