+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 - Ampache API:
   * Action `playlist_songs` returning internal error 500 if the playlist contains any broken track references
 - Song progress shown incorrectly in the media session integration of Chrome when playing (exotic file types) with the fallback Aurora.js player
+- Track disappearing from playlists when moved to another folder within the library folder
+  [#1173](https://github.com/owncloud/music/issues/1173)
 
 ## 2.0.1 - 2024-09-08